fixed oib reader for windows

+++ fluorender/FluoRender/Formats/oib_reader.h
@@ -102,10 +102,10 @@ class OIBReader : public BaseReader
       void ReadSingleOib();
       void ReadSequenceOib();
 #ifdef WIN32
-      void ReadStream(IStorage *pStg, wstring &stream_name);
-      void ReadOibInfo(uint8_t* pbyData, size_t size);
-      void ReadOif(uint8_t* pbyData, size_t size);
-      void ReadTiff(uint8_t* pbyData, unsigned short *val, int z);
+	void ReadStream(IStorage *pStg, wstring &stream_name);
+	void ReadOibInfo(BYTE* pbyData, ULONG size);
+	void ReadOif(BYTE* pbyData, ULONG size);
+	void ReadTiff(BYTE* pbyData, unsigned short *val, int z);
 #endif
 };
